Opening scene, scene transitions, invisible walls
main menu opens to the opening scene. opening scene has some trees -
still needs more design work and assets. if you walk to the right it
switches to police station street scene. all invisible walls are in
place (i think?). if you go into the big building it starts the puzzle.
and if you beat the puzzle it goes back to main menu.