Preliminary sound updates:  READ DESCRIPTION! Beach music and environment noise added, stuff cleaned up and reorganized. :sound:
The 3 bird_river tracks are meant to be played one right after the other
in any order for continuous noise. The flyby is a punctuating sound to
be played at random intervals. IMPORTANT: BECAUSE OF SOUND FILE
REORGANIZATION, MUSIC WILL BE BROKEN. The tracks have simply been
disconnected, there is no bug or audio problem.