Added cave music and ending music, walking on sand sound effects for player added :sound: :musical_note:
I hijacked some of the tracks I made for demos at the beginning of the
semester if we need more music for the ending (if we get that far). The
cave music is just the demo right now, but it still loops and will work
for the final if needed.